+++ src/datasources/asset/constants/ListAssets.constants.ts
@@ -9,6 +9,7 @@ export enum ListAssetsFieldNames {
     VENDOR_NAME = 'VendorName',
     BUS_TYPE = 'BusType',
     ASSET_TYPE = 'AssetType',
+    CALIBRATION_DUE_DATE = 'ExternalCalibration.NextRecommendedDate'
 }
 
 export const ListAssetsFields: Record<string, QBField> = {
@@ -72,6 +73,19 @@ export const ListAssetsFields: Record<string, QBField> = {
             dataSource: AssetTypeOptions,
         },
     },
+    CALIBRATION_DUE_DATE: {
+        label: 'Calibration Due Date',
+        dataField: ListAssetsFieldNames.CALIBRATION_DUE_DATE,
+        filterOperations: [
+            QueryBuilderOperations.LESS_THAN.name,
+            QueryBuilderOperations.LESS_THAN_OR_EQUAL_TO.name,
+            QueryBuilderOperations.GREATER_THAN.name,
+            QueryBuilderOperations.GREATER_THAN_OR_EQUAL_TO.name
+        ],
+        lookup: {
+            dataSource: []
+        }
+    }
 };
 
 export const ListAssetsStaticFields = [