+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## [3.42.0](https://github.com/ni/systemlink-grafana-plugins/compare/v3.41.0...v3.42.0) (2024-10-25)
+
+### Features
+
+* **asset:** more friendly options for calibration due date and add new option ([#106](https://github.com/ni/systemlink-grafana-plugins/issues/106)) ([48fa51c](https://github.com/ni/systemlink-grafana-plugins/commit/48fa51cf332600cdbf718aa3e73d757852a20e5c))
+
+## [3.41.0](https://github.com/ni/systemlink-grafana-plugins/compare/v3.40.1...v3.41.0) (2024-10-25)
+
+### Features
+
+* **asset:** remove includeOnlyDataInTimeRange query param, default true ([#104](https://github.com/ni/systemlink-grafana-plugins/issues/104)) ([29e9eea](https://github.com/ni/systemlink-grafana-plugins/commit/29e9eea9946303e387fd08f06ecb26deb5c79f57))
+
+## [3.40.1](https://github.com/ni/systemlink-grafana-plugins/compare/v3.40.0...v3.40.1) (2024-10-25)
+
+### Bug Fixes
+
+* **asset:** fix data link on deployed env ([#103](https://github.com/ni/systemlink-grafana-plugins/issues/103)) ([f1e4cd2](https://github.com/ni/systemlink-grafana-plugins/commit/f1e4cd2bf29f964ca7f100a7d4fbf7e86c879e4a))
+
+## [3.40.0](https://github.com/ni/systemlink-grafana-plugins/compare/v3.39.0...v3.40.0) (2024-10-24)
+
+### Features
+
+* **asset:** add resolved due date filter for list assets ([#102](https://github.com/ni/systemlink-grafana-plugins/issues/102)) ([d4ad953](https://github.com/ni/systemlink-grafana-plugins/commit/d4ad953970721f1a32c0c944fa1f6c96dae39914))
+
 ## [3.39.0](https://github.com/ni/systemlink-grafana-plugins/compare/v3.38.0...v3.39.0) (2024-10-24)
 
 ### Features